requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
line-bot-sdk==3.11.0
google-generativeai==0.8.3
apscheduler==3.10.4
python-dotenv==1.0.1
httpx==0.27.2
pytz==2024.1
